+++ tools/diagnostics/base_tech_bundle/kafka_bundle.py
@@ -10,6 +10,15 @@ logger = logging.getLogger(__name__)
 
 
 class KafkaBundle(BaseTechBundle):
+
+     def __init__(self, package_name, service_name, scheduler_tasks, service, output_directory):
+        super().__init__(self,
+                         package_name,
+                         service_name,
+                         scheduler_tasks,
+                         service,
+                         output_directory)
+
     def create(self):
         logger.info("Creating Kafka bundle")
         brokers = self.create_broker_list_file()
